+++ crates/k8s-version/src/level/mod.rs
@@ -47,22 +47,18 @@ pub enum Level {
 impl FromStr for Level {
     type Err = ParseLevelError;
 
-    // SAFETY: We purposefully allow the `clippy::unwrap_in_result` lint below in this function.
-    // We can use expect here, because the correct match labels must be used.
-    //
-    // FIXME (@Techassi): This attribute can be used on individual unwrap and expect calls since
-    // Rust 1.91.0. We should move this attribute to not contaminate an unnecessarily large scope
-    // once we bump the toolchain to 1.91.0.
-    // See https://github.com/rust-lang/rust-clippy/pull/15445
-    #[allow(clippy::unwrap_in_result)]
     fn from_str(input: &str) -> Result<Self, Self::Err> {
         let captures = LEVEL_REGEX.captures(input).context(InvalidFormatSnafu)?;
 
+        // SAFETY: We purposefully allow the `clippy::unwrap_in_result` lint here and below.
+        // We can use expect here, because the correct match labels must be used.
+        #[allow(clippy::unwrap_in_result)]
         let identifier = captures
             .name("identifier")
             .expect("internal error: check that the correct match label is specified")
             .as_str();
 
+        #[allow(clippy::unwrap_in_result)]
         let version = captures
             .name("version")
             .expect("internal error: check that the correct match label is specified")